pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["financial_assistant"]

[project]
name = "ai-financial-research-assistant"
version = "0.1.0"
description = "AI agent for financial research using FastAPI and Langgraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.18.4",
  "fastapi[standard]>=0.135.2",
  "langchain>=1.2.12",
  "langchain-community>=0.4.1",
  "langchain-google-genai>=4.2.1",
  "langgraph>=1.1.2",
  "pgvector>=0.4.2",
  "psycopg-pool>=3.3.0",
  "psycopg[binary]>=3.3.3",
  "pwdlib[argon2]>=0.3.0",
  "pydantic-settings>=2.13.1",
  "pyjwt>=2.12.1",
  "pypdf>=6.10.0",
  "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
  "pre-commit>=4.5.1",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-env>=1.6.0",
  "ruff>=0.15.7",
]

[tool.ruff]
line-length = 88
target-version = "py313"
# Specify the source directories to lint and exclude common directories that shouldn't be linted
src = ["financial_assistant"]
extend-exclude = [".venv", "__pycache__", "build", "dist", "alembic"]
# Automatically fix issues where possible
fix = true

[tool.ruff.lint]
select = [
  "E",     # pycodestyle errors (standard formatting errors)
  "W",     # pycodestyle warnings
  "F",     # pyflakes (unused variables, imports, etc.)
  "I",     # isort (automatic import sorting)
  "N",     # pep8-naming (enforces naming conventions)
  "B",     # flake8-bugbear (detects common errors and pitfalls in Python)
  "S",     # flake8-security (finds security issues in code)
  "UP",    # pyupgrade (automatically upgrades syntax for newer Python versions)
  "SIM",   # flake8-simplify (suggests code simplifications)
  "RUF",   # Ruff-specific rules
  "ASYNC", # flake8-async (detects errors in async code - important in FastAPI)
]
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.isort]
# Group imports from our app together, separating them from standard and third-party libraries
known-first-party = ["financial_assistant"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
# Uses existing trailing commas as an indication that short lines should be left separate
skip-magic-trailing-comma = false
# Formats code in docstrings
docstring-code-format = true
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
env = ["ENVIRONMENT=testing"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
  "S101", # asserts
  "S105", # hardcoded password strings
  "S106", # hardcoded password function arguments
  "S107", # hardcoded password function defaults
]
